Сделана кодо генерация

+++ internal/tasks/orm.go
@@ -1,7 +1,7 @@
 package tasks
 
 type Task struct {
-	ID     uint   `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID     int64  `gorm:"primaryKey;autoIncrement" json:"id"`
 	Task   string `json:"task"`
 	Status string `json:"status"`
 }
